Fix bugs in MSE and identity lenses, improve lr lens types

+++ OpticML/src/OpticML/Components.hs
@@ -13,11 +13,11 @@ import OpticML.LensImpl (linear, add)
 seed :: Int
 seed = 384723978470123987
 
-dense :: Num a => (Int, Int) -> Para' (Matrix Float, Matrix Float) ((Matrix a, Matrix a), Matrix a) (Matrix a)
+dense :: Num a => (Int, Int) -> Para' (Matrix Double, Matrix Double) ((Matrix a, Matrix a), Matrix a) (Matrix a)
 dense (a, b) = linearP (a, b) |.| biasP b
 
-linearP :: Num a => (Int, Int) -> Para' (Matrix Float) (Matrix a, Matrix a) (Matrix a)
+linearP :: Num a => (Int, Int) -> Para' (Matrix Double) (Matrix a, Matrix a) (Matrix a)
 linearP (a, b) = Para (fromList b a $ take (a * b) (mkNormals seed)) linear
 
-biasP :: Num a => Int -> Para' (Matrix Float) (Matrix a, Matrix a) (Matrix a)
+biasP :: Num a => Int -> Para' (Matrix Double) (Matrix a, Matrix a) (Matrix a)
 biasP b = Para (zero b 1) add
